ACTIN-759 Move method under ResourceLocator
It is still static though.
