+++ algo/src/test/kotlin/com/hartwig/actin/algo/ckb/serialization/CkbExtendedEvidenceJsonTest.kt
@@ -2,7 +2,7 @@ package com.hartwig.actin.algo.ckb.serialization
 
 import com.hartwig.actin.algo.ckb.json.CkbExtendedEvidenceEntry
 import com.hartwig.actin.algo.ckb.json.CkbExtendedEvidenceTestFactory
-import com.hartwig.actin.testutil.resourceOnClasspath
+import com.hartwig.actin.testutil.ResourceLocator.resourceOnClasspath
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
