+++ clinical/src/test/kotlin/com/hartwig/actin/clinical/feed/emc/TestFeedFactory.kt
@@ -12,7 +12,7 @@ import com.hartwig.actin.clinical.feed.emc.questionnaire.QuestionnaireEntry
 import com.hartwig.actin.clinical.feed.emc.questionnaire.TestQuestionnaireFactory
 import com.hartwig.actin.clinical.feed.emc.surgery.SurgeryEntry
 import com.hartwig.actin.clinical.feed.emc.vitalfunction.VitalFunctionEntry
-import com.hartwig.actin.testutil.resourceOnClasspath
+import com.hartwig.actin.testutil.ResourceLocator.resourceOnClasspath
 import java.time.LocalDate
 import java.time.LocalDateTime
