+++ common/src/test/kotlin/com/hartwig/actin/algo/serialization/TreatmentMatchJsonTest.kt
@@ -8,7 +8,7 @@ import com.hartwig.actin.algo.datamodel.TrialMatch
 import com.hartwig.actin.algo.serialization.TreatmentMatchJson.fromJson
 import com.hartwig.actin.algo.serialization.TreatmentMatchJson.read
 import com.hartwig.actin.algo.serialization.TreatmentMatchJson.toJson
-import com.hartwig.actin.testutil.resourceOnClasspath
+import com.hartwig.actin.testutil.ResourceLocator.resourceOnClasspath
 import java.io.File
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
