+++ common/src/test/kotlin/com/hartwig/actin/testutil/ResourceLocator.kt
@@ -1,7 +1,7 @@
 package com.hartwig.actin.testutil
 
-object ResourceLocator
-
-fun resourceOnClasspath(relativePath: String): String {
-    return ResourceLocator.javaClass.getResource("/" + relativePath.removePrefix("/"))!!.path
+object ResourceLocator {
+    fun resourceOnClasspath(relativePath: String): String {
+        return ResourceLocator.javaClass.getResource("/" + relativePath.removePrefix("/"))!!.path
+    }
 }
