+++ common/src/test/kotlin/com/hartwig/actin/trial/serialization/TrialJsonTest.kt
@@ -1,6 +1,6 @@
 package com.hartwig.actin.trial.serialization
 
-import com.hartwig.actin.testutil.resourceOnClasspath
+import com.hartwig.actin.testutil.ResourceLocator.resourceOnClasspath
 import com.hartwig.actin.trial.datamodel.Cohort
 import com.hartwig.actin.trial.datamodel.Eligibility
 import com.hartwig.actin.trial.datamodel.EligibilityFunction
